feat: add `authorization` to `quotes` controller and service layers

+++ services/quotes.service.ts
@@ -1,3 +1,5 @@
+import createError from 'http-errors';
+
 import { IQuote } from '../models/Quote';
 import quoteRepository from '../repositories/quote.repository';
 
@@ -15,15 +17,35 @@ class QuotesService {
   }
 
   async getQuote(id: string) {
-    return await quoteRepository.getQuote(id);
+    var getQuote = await quoteRepository.getQuote(id);
+
+    if (!getQuote) {
+      throw createError.NotFound('Quote not found.');
+    }
+
+    return getQuote;
   }
 
   async updateQuote(id: string, quote: IQuote) {
+    var getQuote = await this.getQuote(id);
+
+    if (getQuote.userID.toString() !== quote.userID.toString()) {
+      throw createError.Unauthorized();
+    }
+
     return await quoteRepository.updateQuote(id, quote);
   }
 
-  async deleteQuote(id: string): Promise<void | null> {
-    return await quoteRepository.deleteQuote(id);
+  async deleteQuote(id: string, userID: string): Promise<void | null> {
+    var getQuote = await this.getQuote(id);
+
+    if (getQuote?.userID.toString() !== userID) {
+      throw createError.Unauthorized();
+    }
+
+    await quoteRepository.deleteQuote(id);
+
+    return;
   }
 }
